honor the --output-file|-o command line option
find-fixes and order-commits declare an option to direct their
output to a file, but both tools ignore that and just spit out
everything to stdout, instead.

This patch adjusts both tools to honor their command line option,
when selected, as well as it makes order-commits to print out
the not-found commit list to stderr, in order to make it easier
to filter the output, if required.

Signed-off-by: Rafael Aquini <aquini@redhat.com>
Signed-off-by: Jeff Moyer <jmoyer@redhat.com>