find-fixes: don't bail on invalid commit hashes
When adding a commit, tolerate non-existent hashes.  Those may be hashes
from maintainer trees and the like.  Also make the warning clearer when
the history walk did not find some commits.  We want to make it obvious
that this is completely unexpected behavior, and that the results may
be incomplete.

Signed-off-by: Jeff Moyer <jmoyer@redhat.com>